pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.2",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
